+++ src/frequenz/channels/__init__.py
@@ -3,15 +3,18 @@
 
 """Channel implementations."""
 
-from frequenz.channels.anycast import Anycast
-from frequenz.channels.base_classes import BufferedReceiver, Peekable, Receiver, Sender
-from frequenz.channels.bidirectional import Bidirectional, BidirectionalHandle
-from frequenz.channels.broadcast import Broadcast
-from frequenz.channels.merge import Merge
-from frequenz.channels.merge_named import MergeNamed
-from frequenz.channels.select import Select
-from frequenz.channels.utils.file_watcher import FileWatcher
-from frequenz.channels.utils.timer import Timer
+from . import util
+from ._anycast import Anycast
+from ._base_classes import (
+    BufferedReceiver,
+    ChannelClosedError,
+    ChannelError,
+    Peekable,
+    Receiver,
+    Sender,
+)
+from ._bidirectional import Bidirectional, BidirectionalHandle
+from ._broadcast import Broadcast
 
 __all__ = [
     "Anycast",
@@ -19,12 +22,10 @@ __all__ = [
     "BidirectionalHandle",
     "Broadcast",
     "BufferedReceiver",
-    "FileWatcher",
-    "Merge",
-    "MergeNamed",
+    "ChannelClosedError",
+    "ChannelError",
     "Peekable",
     "Receiver",
-    "Select",
     "Sender",
-    "Timer",
+    "util",
 ]