+++ src/frequenz/channels/util/__init__.py
@@ -1,0 +1,18 @@
+# License: MIT
+# Copyright © 2022 Frequenz Energy-as-a-Service GmbH
+
+"""Channel utilities."""
+
+from ._file_watcher import FileWatcher
+from ._merge import Merge
+from ._merge_named import MergeNamed
+from ._select import Select
+from ._timer import Timer
+
+__all__ = [
+    "FileWatcher",
+    "Merge",
+    "MergeNamed",
+    "Select",
+    "Timer",
+]