+++ src/frequenz/channels/util/_merge.py
@@ -1,30 +1,43 @@
 # License: MIT
 # Copyright © 2022 Frequenz Energy-as-a-Service GmbH
 
-"""Merge messages coming from channels into a single stream containing name of message."""
+"""Merge messages coming from channels into a single stream."""
 
 import asyncio
 from collections import deque
-from typing import Any, Deque, Set, Tuple
+from typing import Any, Deque, Set
 
-from frequenz.channels.base_classes import ChannelClosedError, Receiver, T
+from .._base_classes import ChannelClosedError, Receiver, T
 
 
-class MergeNamed(Receiver[Tuple[str, T]]):
-    """Merge messages coming from multiple named channels into a single stream."""
+class Merge(Receiver[T]):
+    """Merge messages coming from multiple channels into a single stream.
 
-    def __init__(self, **kwargs: Receiver[T]) -> None:
-        """Create a `MergeNamed` instance.
+    Example:
+        For example, if there are two channel receivers with the same type,
+        they can be awaited together, and their results merged into a single
+        stream, by using `Merge` like this:
+
+        ```python
+        merge = Merge(receiver1, receiver2)
+        while msg := await merge.receive():
+            # do something with msg
+            pass
+        ```
+    """
+
+    def __init__(self, *args: Receiver[T]) -> None:
+        """Create a `Merge` instance.
 
         Args:
-            **kwargs: sequence of channel receivers.
+            *args: sequence of channel receivers.
         """
-        self._receivers = kwargs
+        self._receivers = {str(id): recv for id, recv in enumerate(args)}
         self._pending: Set[asyncio.Task[Any]] = {
             asyncio.create_task(recv.__anext__(), name=name)
             for name, recv in self._receivers.items()
         }
-        self._results: Deque[Tuple[str, T]] = deque(maxlen=len(self._receivers))
+        self._results: Deque[T] = deque(maxlen=len(self._receivers))
 
     def __del__(self) -> None:
         """Cleanup any pending tasks."""
@@ -32,10 +45,10 @@ class MergeNamed(Receiver[Tuple[str, T]]):
             task.cancel()
 
     async def ready(self) -> None:
-        """Wait until there's a message in any of the channels.
+        """Wait until the receiver is ready with a value.
 
         Raises:
-            ChannelClosedError: when all the channels are closed.
+            ChannelClosedError: if the underlying channel is closed.
         """
         # we use a while loop to continue to wait for new data, in case the
         # previous `wait` completed because a channel was closed.
@@ -55,20 +68,20 @@ class MergeNamed(Receiver[Tuple[str, T]]):
                 if isinstance(item.exception(), StopAsyncIteration):
                     continue
                 result = item.result()
-                self._results.append((name, result))
+                self._results.append(result)
                 self._pending.add(
                     # pylint: disable=unnecessary-dunder-call
                     asyncio.create_task(self._receivers[name].__anext__(), name=name)
                 )
 
-    def consume(self) -> Tuple[str, T]:
+    def consume(self) -> T:
         """Return the latest value once `ready` is complete.
 
         Raises:
             EOFError: When called before a call to `ready()` finishes.
 
         Returns:
-            The next value that was received, along with its name.
+            The next value that was received.
         """
         assert self._results, "calls to `consume()` must be follow a call to `ready()`"