+++ src/frequenz/channels/util/_merge_named.py
@@ -1,43 +1,30 @@
 # License: MIT
 # Copyright © 2022 Frequenz Energy-as-a-Service GmbH
 
-"""Merge messages coming from channels into a single stream."""
+"""Merge messages coming from channels into a single stream containing name of message."""
 
 import asyncio
 from collections import deque
-from typing import Any, Deque, Set
+from typing import Any, Deque, Set, Tuple
 
-from frequenz.channels.base_classes import ChannelClosedError, Receiver, T
+from .._base_classes import ChannelClosedError, Receiver, T
 
 
-class Merge(Receiver[T]):
-    """Merge messages coming from multiple channels into a single stream.
+class MergeNamed(Receiver[Tuple[str, T]]):
+    """Merge messages coming from multiple named channels into a single stream."""
 
-    Example:
-        For example, if there are two channel receivers with the same type,
-        they can be awaited together, and their results merged into a single
-        stream, by using `Merge` like this:
-
-        ```python
-        merge = Merge(receiver1, receiver2)
-        while msg := await merge.receive():
-            # do something with msg
-            pass
-        ```
-    """
-
-    def __init__(self, *args: Receiver[T]) -> None:
-        """Create a `Merge` instance.
+    def __init__(self, **kwargs: Receiver[T]) -> None:
+        """Create a `MergeNamed` instance.
 
         Args:
-            *args: sequence of channel receivers.
+            **kwargs: sequence of channel receivers.
         """
-        self._receivers = {str(id): recv for id, recv in enumerate(args)}
+        self._receivers = kwargs
         self._pending: Set[asyncio.Task[Any]] = {
             asyncio.create_task(recv.__anext__(), name=name)
             for name, recv in self._receivers.items()
         }
-        self._results: Deque[T] = deque(maxlen=len(self._receivers))
+        self._results: Deque[Tuple[str, T]] = deque(maxlen=len(self._receivers))
 
     def __del__(self) -> None:
         """Cleanup any pending tasks."""
@@ -45,10 +32,10 @@ class Merge(Receiver[T]):
             task.cancel()
 
     async def ready(self) -> None:
-        """Wait until the receiver is ready with a value.
+        """Wait until there's a message in any of the channels.
 
         Raises:
-            ChannelClosedError: if the underlying channel is closed.
+            ChannelClosedError: when all the channels are closed.
         """
         # we use a while loop to continue to wait for new data, in case the
         # previous `wait` completed because a channel was closed.
@@ -68,20 +55,20 @@ class Merge(Receiver[T]):
                 if isinstance(item.exception(), StopAsyncIteration):
                     continue
                 result = item.result()
-                self._results.append(result)
+                self._results.append((name, result))
                 self._pending.add(
                     # pylint: disable=unnecessary-dunder-call
                     asyncio.create_task(self._receivers[name].__anext__(), name=name)
                 )
 
-    def consume(self) -> T:
+    def consume(self) -> Tuple[str, T]:
         """Return the latest value once `ready` is complete.
 
         Raises:
             EOFError: When called before a call to `ready()` finishes.
 
         Returns:
-            The next value that was received.
+            The next value that was received, along with its name.
         """
         assert self._results, "calls to `consume()` must be follow a call to `ready()`"