+++ RELEASE_NOTES.md
@@ -18,6 +18,26 @@ time.
   `new_receiver()` and `new_sender()` respectively. This is to make it more
   clear that new objects are being created.
 
+* The public API surface has been reduced considerably to make it more clear
+  where to import symbols.  You should update your imports.  The new symbol
+  locations are:
+
+  * `frequenz.channels.Anycast`
+  * `frequenz.channels.Broadcast`
+  * `frequenz.channels.Anycast`
+  * `frequenz.channels.Bidirectional`
+  * `frequenz.channels.BidirectionalHandle`
+  * `frequenz.channels.Broadcast`
+  * `frequenz.channels.BufferedReceiver`
+  * `frequenz.channels.Peekable`
+  * `frequenz.channels.Receiver`
+  * `frequenz.channels.Sender`
+  * `frequenz.channels.util.Merge`
+  * `frequenz.channels.util.MergeNamed`
+  * `frequenz.channels.util.FileWatcher`
+  * `frequenz.channels.util.Select`
+  * `frequenz.channels.util.Timer`
+
 ## New Features
 
 <!-- Here goes the main new features and examples or instructions on how to use them -->